scripts/publish: Don't support patch/minor/major
Require that the developer provide a version string, rather than relying
on npm/yarn to do the right thing re: semantic versioning. Since we're
still in beta, it's easier that we manage our own version strings so
that, for example, `v0.0.1-beta.9` doesn't get revved up to `v0.0.1`
when passing `patch` as an argument.